requirements.txt
# Development & Linting
ruff
pytest
pytest-dotenv
python-dotenv

# Core & Data Models
pydantic

# Vector Database & Math
# FAISS GPU: Installed via conda (no Windows pip wheels available)
# Setup: conda install -c conda-forge faiss-gpu (requires Python 3.12)
# The venv-win environment is a conda env with faiss-gpu pre-installed
numpy

# Embedding Backends
voyageai>=0.2.0
sentence-transformers>=2.2

# Data Fetching & Datasets
datasets
Pillow
tqdm

# LLM Backends
openai>=1.0.0

# HTTP Client (for Kroki rendering)
httpx>=0.25.0

